requirements.txt
streamlit>=1.30
numpy>=1.24
faiss-cpu>=1.7.4
sentence-transformers>=2.2
requests>=2.31